# packages/types-app/openapi/platform.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "familiar.systems platform API",
    "description": "Auth, campaign CRUD, routing table, and shard coordination.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/campaigns": {
      "get": {
        "tags": ["campaigns"],
        "summary": "`GET /api/campaigns` — list the authenticated user's campaigns,\nmost-recent first. Reads only the platform-side routing table; no\nfan-out to shards.",
        "operationId": "list_campaigns",
        "responses": {
          "200": {
            "description": "List of campaigns",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Campaign"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["campaigns"],
        "summary": "`POST /api/campaigns` — mint a CampaignId, ask the campaign tier to\ninitialize per-campaign state, write the routing row, return the id.",
        "description": "Idempotent on `idempotency_token`: a retry with the same token returns\nthe same `campaign_id`. The order (write `create_attempts` first, then\ncall shard, then write `campaigns`) is what makes retries safe — see\nthe long form in the design doc.",
        "operationId": "create_campaign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Campaign created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCampaignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Shard or DB failure"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/health": {
      "get": {
        "tags": ["system"],
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Service is healthy"
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": ["auth"],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Campaign": {
        "type": "object",
        "description": "One row in the response of `GET /api/campaigns`.\n\nMirrored from the campaign tier on every successful initialize and\nsettings edit. Fields marked `Option` are `None` until the wizard's\nSeal call commits, after which they are mirrored over.\n\n`last_init_error`: populated by the campaign tier via\n`POST /internal/platform/campaigns/<id>/init-failed` when an initialize\nattempt fails. Distinct from \"no init attempt yet\" (`wizard_completed_at IS NULL`\nAND `last_init_error IS NULL`).",
        "required": ["id", "created_at", "updated_at"],
        "properties": {
          "content_locale": {
            "type": ["string", "null"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "game_system": {
            "type": ["string", "null"]
          },
          "id": {
            "$ref": "#/components/schemas/CampaignId"
          },
          "last_init_error": {
            "type": ["string", "null"]
          },
          "name": {
            "type": ["string", "null"]
          },
          "tagline": {
            "type": ["string", "null"]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "wizard_completed_at": {
            "type": ["string", "null"],
            "format": "date-time"
          }
        }
      },
      "CampaignId": {
        "type": "string",
        "format": "nanoid",
        "description": "Identifies a campaign. Kept as a nanoid for short URLs."
      },
      "CreateCampaignRequest": {
        "type": "object",
        "description": "Body for `POST /api/campaigns`.\n\nThe SPA mints `idempotency_token` (a fresh nanoid per click) so that a\nretry of the same logical create call returns the same `CampaignId`\nrather than allocating a new one.",
        "required": ["idempotency_token"],
        "properties": {
          "idempotency_token": {
            "type": "string"
          }
        }
      },
      "CreateCampaignResponse": {
        "type": "object",
        "description": "Response body for `POST /api/campaigns`.",
        "required": ["campaign_id"],
        "properties": {
          "campaign_id": {
            "$ref": "#/components/schemas/CampaignId"
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "description": "Response body for `GET /me`.\n\n**Wire format we own (API contract with the SPA).** Exported to TypeScript\nas `MeResponse` via ts-rs. Changes are breaking changes to the frontend.\n\n`id` is the Hanko subject (= `users.id`); see\n`apps/platform/src/entities/users.rs`.",
        "required": ["id", "email"],
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/UserId"
          }
        }
      },
      "UserId": {
        "type": "string",
        "format": "uuid",
        "description": "Uniquely identifies a user."
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Hanko session token"
      }
    }
  }
}