# packages/types-app/src/generated/campaigns/CreateCampaignResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CampaignId } from "../id/CampaignId";

/**
 * Response body for `POST /api/campaigns`.
 */
export type CreateCampaignResponse = { campaign_id: CampaignId };